Update Property_price_pred_code.R
 Updated a day ago
Problem at hand: Everyone has a dream to “own a house” and being conscious about the price of the property is very important when buying a new one. Multiple information sources like websites, reality tv shows and dealers/brokers create a lot of confusion.

Project objective: What if, a machine estimates the price based on the characteristics of the property! Project Objective is to identify the relationship between characteristics of the property and price of the property and identify best approach to predict the price of the property.

Data information: Second-hand cross-sectional data about the properties sold in Iowa from 2006 to 2010 Data has information about 2930 properties Data has 80 different variables that define different physical attributes of the property

Caution: Data is about the price of the properties in Iowa and since the price of the property changes from state to state, our predictions might be valid only for estimating the property prices in Iowa Data has information only till 2010 and for adjusting to present scenario “change in property price from 2010 to present” can be used as adjustment factor