Add files via upload
What will be the death patterns in 2016 and 2017. A deep ARIMA forecasting has been performed in R studio to understand the death patterns of each state over the years, understanding what the future shall be? This panel data set has been corrected for stationarity, auto correlation to ensure that the forecasting is accurately done.  Further, we have deep dived into the search patterns of people in different states to understand what are the most common searches and which ones contribute to deaths in that state. 
You could use filters to: 
1. To breakdown the forecast by death type 
2. To understand the death rate in the future for different states 
3. To understand which are the search terms that are associated with deaths in different states by reading the P values of the terms which have been used to construct the regression line