Create Syngenta Crop Analytics
Quantifying the crop yield is essential to make most effective use of policies to ensure food security.  This paper aims at providing quantitative framework for predicting corn hybrid performance in new, untested locations which can help agriculture industry breed better seeds and faster based on its genetics and environmental factors. We tried to develop models that identify robust patterns in corn hybrid genetics in conjunction with environmental data such as soil and weather that may help scientists more accurately choose seeds that increase the productivity of the crops we plant in the untested locations – and help address the growing global food demand. So, we extracted the features from the three separate datasets such as weather, soil and genetics and yield. We clustered environmental factors into 6 groups based on its properties. Next, we analyzed 2267 corn hybrids in 2122 locations between the year 2008 to 2016 and built the model for predicting corn performance for each cluster.
According to the rules of Syngenta Crop Challenge, for which this research was conducted, we predicted yield using artificial neural network for untested locations across clusters of 6 sub regions and 5 corn hybrid varieties clusters which we derived using K-means Clustering for environment and hybrid variables