Create Sales prediction (Home products Retailer)
The retail enterprise can no longer rely on just Brick and Mortar stores to succeed in a digital world. A powerful online presence is key for any retailer to succeed. The online platform presents a key place for retailers to sell their product. It is a sector that has immense growth year-over-year. Apart from being a market place it is also a tool that can be used to increase footfall into the stores and drive business. 
Keeping both these aspects in mind and the retailers business of being a strong online and brick and mortar retailer of Home and Lifestyle goods we have created a deep-dive analysis tool which compares online and offline aspects of the retailer’s  Sales and will enables its user to slice and dice data at various levels to understand subtle nuances that affect sales.